requirements.txt
flask_sqlalchemy==2.1.0
flask-restful==0.3.6
flask-cors==3.0.7
flask-cache==0.13.1
flask-marshmallow==0.9.0
